// front/src/models/bills/BillStatus.ts
export enum BillStatus {
  NONE = 'NONE',
  REMOVED = 'REMOVED',
  REJECTED = 'REJECTED',
  EXPIRED = 'EXPIRED',
  PENDING_PAY = 'PENDING_PAY',
  PAYED = 'PAYED',
  PENDING_PROCESSING = 'PENDING_PROCESSING',
  PROCESSING = 'PROCESSING',
  SENT = 'SENT',
  AWAIT_DELIVERY = 'AWAIT_DELIVERY',
  DELIVERED = 'DELIVERED'
}